fix: move util.get_from_db to repositories.user
