+++ src/services/scsc.py
@@ -17,10 +17,10 @@ from src.repositories import (
     SigRepositoryDep,
     StandbyReqTblRepositoryDep,
     UserRepositoryDep,
+    get_user_role_level,
 )
 from src.util import (
     get_new_year_semester,
-    get_user_role_level,
     map_semester_name,
     utcnow,
 )
