+++ src/services/user.py
@@ -18,12 +18,12 @@ from src.repositories import (
     StandbyReqTblRepositoryDep,
     UserRepositoryDep,
     UserRoleRepositoryDep,
+    get_user_role_level,
 )
 from src.schemas import PublicUserResponse, UserResponse
 from src.util import (
     DepositDTO,
     generate_user_hash,
-    get_user_role_level,
     is_valid_img_url,
     is_valid_phone,
     is_valid_student_id,
